refactor(cockpit_ui): extract inbox row-format helpers (refs #1354)
Move the inbox row-formatting + plan-review helpers — 36 names spanning
sort keys, row renderers (regular, plan-review decision card, reply,
thread dispatcher), triage helpers, message-detail block builders, PM
target / persona / primer resolvers, plan-review label & content
plumbing, and the lens taxonomy — into a new
``pollypm.cockpit_ui_inbox_format`` module. ``cockpit_ui`` re-exports
the names so existing imports and
``monkeypatch.setattr("pollypm.cockpit_ui._format_inbox_thread_row", ...)``
test hooks keep working.

Second slice of the rolling ``cockpit_ui.py`` god-module split
tracked by #1354 (first slice: PR #1911, ``cockpit_settings_gather``).
Pure module-boundary move: no behavior changes, no public-API churn
beyond the test-patchable symbols still being re-exported from the
original location. ``cockpit_ui.py`` shrinks 17,838 to 16,798 LOC
(-1,040); the new module is 1,165 LOC and ruff-clean. Remaining
cockpit_ui.py ruff errors dropped from 50 to 49 (the moved code's
``rejection_feedback`` imports are gone).

- src/pollypm/cockpit_ui_inbox_format.py: new module with the inbox
  sort key + ISO-weight helpers, sender / triage / archive-message
  formatters, user-prompt + heuristic-action + triage-banner block
  renderers, ``_format_inbox_row`` / ``_format_inbox_plan_review_row``
  / ``_format_inbox_reply_row`` / ``_format_inbox_thread_row``,
  rollup / fuzzy-match / recent-timestamp predicates, project-PM
  persona+label resolvers, ``_resolve_pm_target`` /
  ``_build_pm_context_line`` PM input plumbing,
  ``_extract_plan_review_meta`` / ``_plan_content_for_review`` /
  ``_extract_plan_summary_block`` / ``_extract_plan_judgment_calls``
  / ``_extract_blocking_question_meta`` plan-review label parsers,
  ``_plan_review_has_round_trip`` + ``_build_plan_review_primer`` /
  ``_build_plan_review_denial_primer`` PM-primer builders,
  ``_extract_proposal_spec``, and the ``_INBOX_LENSES`` /
  ``_INBOX_LENS_KINDS`` / ``_INBOX_DEFAULT_LENS_SLUG`` /
  ``_INBOX_PRIORITY_RANK`` lens constants.
  ``_render_heuristic_action_block`` and ``_plan_content_for_review``
  lazy-import the dashboard-body / plan-path helpers from
  ``cockpit_ui`` to avoid a module-level cycle while #1354 is in
  flight.
- src/pollypm/cockpit_ui.py: drop the moved block (lines 5480-6557);
  replace with a single re-export
  ``from pollypm.cockpit_ui_inbox_format import (...)``. Trim the
  now-unused ``feedback_target_task_id`` / ``is_rejection_feedback_task``
  imports (their only call sites moved to the new module).
